# likes/docs/swagger.yaml
definitions:
  model.Media:
    properties:
      content:
        type: string
      created_at:
        type: string
      media:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: This is API for likes service
  title: Likes API documentation
  version: 1.0.0
paths:
  /delete:
    delete:
      description: Delete user
      parameters:
      - description: User id
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /delete_tweet:
    delete:
      description: Delete by tweet_id
      parameters:
      - description: Tweet ID
        in: query
        name: tweet_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /home_timeline:
    get:
      description: Retrieve home timeline
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Media'
            type: array
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /like_tweet:
    post:
      description: Like specific tweet
      parameters:
      - description: User ID
        in: body
        name: user_id
        required: true
        schema:
          type: integer
      - description: Tweet ID
        in: body
        name: tweet_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /login:
    post:
      description: Login for user
      parameters:
      - description: Password
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: Email
        in: body
        name: email
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /post_tweet:
    post:
      description: Post tweet
      parameters:
      - description: User ID
        in: body
        name: user_id
        required: true
        schema:
          type: integer
      - description: Content
        in: body
        name: content
        required: true
        schema:
          type: string
      - description: Retweet ID
        in: body
        name: retweet_id
        schema:
          type: integer
      - description: Media
        in: formData
        name: media
        type: file
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /register:
    post:
      description: Register new user
      parameters:
      - description: Password
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: Email
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: Nickname
        in: body
        name: nickname
        required: true
        schema:
          type: string
      - description: First name
        in: body
        name: first_name
        schema:
          type: string
      - description: Last name
        in: body
        name: last_name
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /retrieve_tweet:
    get:
      description: Retrieve either by tweet_id or user_id
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: integer
      - description: Tweet ID
        in: query
        name: tweet_id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Media'
            type: array
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /tweets_user:
    get:
      description: Retrieve all tweet liked by user
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /unlike_tweet:
    delete:
      description: Unlike specific tweet
      parameters:
      - description: User ID
        in: body
        name: user_id
        required: true
        schema:
          type: integer
      - description: Tweet ID
        in: body
        name: tweet_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /update:
    put:
      description: Update user data
      parameters:
      - description: User id
        in: body
        name: user_id
        required: true
        schema:
          type: integer
      - description: Nickname
        in: body
        name: nickname
        schema:
          type: string
      - description: First name
        in: body
        name: first_name
        schema:
          type: string
      - description: Last name
        in: body
        name: last_name
        schema:
          type: string
      - description: Email
        in: body
        name: email
        schema:
          type: string
      - description: Password
        in: body
        name: password
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
  /users_tweet:
    get:
      description: Retrieve all users who liked tweet
      parameters:
      - description: Tweet ID
        in: query
        name: tweet_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            type: integer
        "405":
          description: Method Not Allowed
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
swagger: "2.0"